refactor: move styles into components files
