+++ src/screens/Login.tsx
@@ -1,6 +1,6 @@
-import { SafeAreaView, Text, View } from 'react-native';
+import { SafeAreaView, StyleSheet, Text, View } from 'react-native';
 import React from 'react';
-import AppButton from '../../components/AppButton';
+import AppButton from '../components/AppButton';
 
 const Login = () => {
   return (
@@ -14,3 +14,8 @@ const Login = () => {
 };
 
 export default Login;
+
+const styles = StyleSheet.create({
+
+});
+
