--- src/screens/Login/styles.js
@@ -1,5 +1,0 @@
-import { StyleSheet } from "react-native";
-
-const styles = StyleSheet.create({
-
-});
